requirements.txt
pyyaml
matplotlib
ruff
docformatter
pydocstyle
mypy
vulture
radon
pytest
